+++ crates/client/rpc/src/lib.rs
@@ -199,7 +199,6 @@ where
 /// Taken from https://github.com/paritytech/substrate/blob/master/client/rpc/src/author/mod.rs#L78
 const TX_SOURCE: TransactionSource = TransactionSource::External;
 
-#[allow(unused_variables)]
 impl<A, B, BE, G, C, P, H> MadaraRpcApiServer for Starknet<A, B, BE, G, C, P, H>
 where
     A: ChainApi<Block = B> + 'static,
@@ -221,8 +220,7 @@ where
             .predeployed_accounts
             .into_iter()
             .map(|account| {
-                let contract_address: FieldElement = Felt252Wrapper(account.contract_address).into();
-                let class_hash: FieldElement = Felt252Wrapper(account.class_hash).into();
+                let contract_address: FieldElement = account.contract_address.into();
                 let balance_string = &self
                     .call(
                         FunctionCall {
@@ -243,7 +241,6 @@ where
 }
 
 #[async_trait]
-#[allow(unused_variables)]
 impl<A, B, BE, G, C, P, H> StarknetWriteRpcApiServer for Starknet<A, B, BE, G, C, P, H>
 where
     A: ChainApi<Block = B> + 'static,